==> step 2_Fancyimpute.ipynb <==
# %load test_fancy.py
#from fancyimpute import KNN
#from sklearn.neighbors import KNeighborsClassifier
#from fancyimpute import BiScaler, KNN, NuclearNormMinimization, SoftImpute
#import numpy as np
#mport statsmodels.imputation.mice as MICE
#df = pd.read_csv("PartitiallyImputedData.csv")
#array = np.array([['a', 'a', 'd', np.nan], ['c', 'c', np.nan, 'f']])
#df = pd.DataFrame(array, columns = ['A','B','C','D'])
#df["A"] = df["A"].astype('category')
#df["B"] = df["B"].astype('category')
#df["C"] = df["C"].astype('category')
#df["D"] = df["D"].astype('category')
#imp = MICE.MICEData(df)


from fancyimpute import IterativeImputer

#from fancyimpute import KNN
import pandas as pd

df = pd.read_csv('miceData_category.csv')
col = df.columns
imputer = IterativeImputer(sample_posterior=True)
data_impute= imputer.fit_transform(df.values)

#X_filled_knn = KNN(k=3).fit_transform(df.values)
df_final = pd.DataFrame(data=data_impute, columns=col)

df_final.to_csv('Iterative_sample_posteriorTrue.csv')

#ii_mse = ((X_filled_ii[missing_mask] - X[missing_mask]) ** 2).mean()
#print("Iterative Imputer norm minimization MSE: %f" % ii_mse)

#knn_mse = ((X_filled_knn[missing_mask] - X[missing_mask]) ** 2).mean()
#print("knnImpute MSE: %f" % knn_mse)
