==> step 3.0 Ordinal_Nominal_CateFeatures_OneHotEncoding.ipynb <==
# check imputatopm works

import pandas as pd

data = pd.read_csv('Iterative_sample_posteriorTrue.csv')
data = data.loc[:, ~data.columns.str.contains('^Unnamed')]

# features applied Multivariate Imputation

col = ['emp_length','mths_since_last_delinq','mths_since_last_record','next_pymnt_d','mths_since_last_major_derog',
       'mths_since_rcnt_il','il_util', 'bc_open_to_buy', 'bc_util', 'mo_sin_old_il_acct','mths_since_recent_bc',
       'mths_since_recent_bc_dlq', 'mths_since_recent_inq', 'mths_since_recent_revol_delinq','percent_bc_gt_75']

# check no missing values

print(data.columns[data.isna().sum() > 0])

# check no float numbers in categorical values

for i in col:
    print(pd.api.types.is_float(i))

print(list(data.columns))

data.drop('zip_code',inplace=True,axis=1)

# categorical feature list without loan_status

categorical_list = ['term',
 'int_rate',
 'grade',
 'sub_grade',
 'emp_length',
 'home_ownership',
 'verification_status',
 'issue_d',
 'purpose',
 'title',
 'addr_state',
 'loan_status',
 'earliest_cr_line',
 'initial_list_status',
 'last_pymnt_d',
 'next_pymnt_d',
 'last_credit_pull_d',
 'application_type',
 'verification_status_joint',
 'sec_app_earliest_cr_line',
 'hardship_status',
 'hardship_loan_status',
 'disbursement_method',
 'debt_settlement_flag',
 'settlement_status']

#哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈哈
# obj_df = data.select_dtypes(include=['category']).copy()
# obj_df.head()

data.groupby(by=['loan_status']).size()

replace_map = {'loan_status':{"0.0": "Charged Off", "1.0": "Current", "2.0": "Default", "3.0": "Fully Paid", "4.0": "In Grace Period", 
                              "5.0": "Late (16-30 days)", "6.0": "Late (31-120 days)"}}

data['loan_status'].replace(0.0, "Charged Off", inplace=True)
data['loan_status'].replace(1.0, "Current", inplace=True)
data['loan_status'].replace(2.0, "Default", inplace=True)
data['loan_status'].replace(3.0, "Fully Paid", inplace=True)
data['loan_status'].replace(4.0, "In Grace Period", inplace=True)
data['loan_status'].replace(5.0, "Late (16-30 days)", inplace=True)
data['loan_status'].replace(6.0, "Late (31-120 days)", inplace=True)

data.groupby(by=['loan_status']).size()

category = []
header = list(data)
for i in header:
    if data[i].dtype.name == 'category':
        category.append(i)

data.groupby(by=['emp_length']).size()

data['emp_length'].replace(0.0, 11.0, inplace=True)
data['emp_length'].replace(1.0, 12, inplace=True)
data['emp_length'].replace(10.0, 13, inplace=True)

data['emp_length'].replace(11.0, 1.0, inplace=True)
data['emp_length'].replace(12.0, 10.0, inplace=True)
data['emp_length'].replace(13.0, 0.0, inplace=True)

data.groupby(by=['emp_length']).size()

data.shape

# drop NA > 50% in original dataset
data.drop(columns=['mths_since_last_delinq', 'mths_since_last_record', 'mths_since_last_major_derog', 
                   'mths_since_recent_bc_dlq', 'mths_since_recent_revol_delinq'], axis=1, inplace=True)

data.groupby(by=['next_pymnt_d']).size()

data['next_pymnt_d'].replace(0.0, 'Apr-2019', inplace=True)
data['next_pymnt_d'].replace(1.0, 'Feb-2019', inplace=True)
data['next_pymnt_d'].replace(2.0, 'Mar-2019', inplace=True)
data['next_pymnt_d'].replace(-1.0, 'Unkown', inplace=True)

nominal_data = pd.read_csv('nominal_data.csv')
nominal_data = nominal_data.loc[:, ~nominal_data.columns.str.contains('^Unnamed')]
nominal_data.drop('next_pymnt_d',axis=1, inplace=True)

nominal_data

ordinal = ['int_rate', 'emp_length', 'earliest_cr_line', 'sec_app_earliest_cr_line']
nominal = ['term',
 'grade',
 'sub_grade',
 'home_ownership',
 'verification_status',
 'issue_d',
 'purpose',
 'title',
 'addr_state',
 'initial_list_status',
 'last_pymnt_d',
 'last_credit_pull_d',
 'application_type',
 'verification_status_joint',
 'hardship_status',
 'hardship_loan_status',
 'disbursement_method',
 'debt_settlement_flag',
 'settlement_status']

data.drop(columns=nominal, axis=1, inplace=True)

data = pd.concat([data, nominal_data], axis=1)

# one-hot encoding

data = pd.get_dummies(data, columns=nominal, prefix=nominal)
data = pd.get_dummies(data, columns=['next_pymnt_d'], prefix=['next_pymnt_d'])

data.shape

data.to_csv('one_hot_encoded_data.csv')

for i in list(data.columns):
    print(data[i].dtype)

data_target = data['loan_status']
data.drop('loan_status',inplace=True, axis=1)
data_input = data
