==> step 4 RandomForest_PCA.ipynb <==
import pandas as pd

data = pd.read_csv('../one_hot_encoded_data.csv')
data = data.loc[:, ~data.columns.str.contains('^Unnamed')]

data.head()

# convert one-hot encoded features to float type

for i in list(data.columns):
    if data[i].dtype == 'int64':
        data[i] = data[i].astype('float')

data_target = data['loan_status']
data.drop('loan_status', inplace=True, axis=1)
data_input = data

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score
#from sklearn.model_selection import KFold

rf = RandomForestClassifier()
score = cross_val_score(rf, data, data_target, cv = 3).mean()
print(score)

from sklearn.preprocessing import StandardScaler
#from sklearn.model_selection import train_test_split
#X_train, X_test, y_train, y_test = train_test_split(data, data_target, test_size=0.2, random_state=42)

X = data
ss = StandardScaler()
X_scaled = ss.fit_transform(X)

rf_1 = RandomForestClassifier()
score_1 = cross_val_score(rf_1, X_scaled, data_target, cv = 3).mean()
print(score_1)

from sklearn.decomposition import PCA

n_components = []
for i in range(80):
    if i*3 < 267:
        n = 2 + i*3
        n_components.append(n)
print(n_components)

from sklearn.model_selection import cross_validate
import math

results = []
for i in n_components:
    pca = PCA(n_components=i)
    x_pca = pca.fit_transform(data)
    rf = RandomForestClassifier()
    cv = cross_validate(rf, x_pca, data_target, cv=3)
    results.append(cv['test_score'].mean())

results.index(max(results))

n_components[57]

max(results)

results

from sklearn.metrics import recall_score
from sklearn.model_selection import train_test_split

pca = PCA(n_components=173)
x_pca = pca.fit_transform(data)

X_train, X_test, y_train, y_test = train_test_split(x_pca, data_target, test_size=0.3, random_state=42)
clf = RandomForestClassifier()
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
recall_score(y_test, y_pred, average=None)

recall_score(y_test, y_pred, average='macro')

recall_score(y_test, y_pred, average='weighted')

recall_score(y_test, y_pred, average='micro') 

from sklearn.metrics import classification_report
print(classification_report(y_test, y_pred, target_names=data_target))

results_cv_173 = []
pca = PCA(n_components=173)
x_pca = pca.fit_transform(data)

for i in solver:
    for j in shrinkage:
        clf = LinearDiscriminantAnalysis(solver=i, shrinkage=j)
        cv_173 = cross_validate(clf, X_pca173, data_target, cv=3)
        cv_8 = cross_validate(clf, X_pca8, data_target, cv=3)
        results_cv_209.append(cv_209['test_score'].mean())
