==> step 4 XGBoost.ipynb <==
import pandas as pd

data = pd.read_csv("../one_hot_encoded_data.csv", index_col=[0])

data.head()

X = data.drop(columns=["loan_status"])

data["loan_status"] = data["loan_status"].astype("category")
category = data["loan_status"].cat.categories

y = data["loan_status"].cat.codes

y.unique()

category
category[y]

#### Xgboost manages only numeric vectors.
https://xgboost.readthedocs.io/en/latest/R-package/discoverYourData.html


import xgboost as xgb
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score



X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=666)



# Xgboost manages only numeric vectors.
dtrain = xgb.DMatrix(X_train.values, label=y_train.values)
dtest = xgb.DMatrix(X_test.values, label=y_test.values)

params = {
    'booster': 'gbtree',
    'objective': 'multi:softmax',
    'num_class': 7,
    'gamma': 0.1, 
    'max_depth': 6,
    'lambda': 2,
    'subsample': 0.7,
    'colsample_bytree': 0.7,
    'min_child_weight': 3,
    'silent': 1,
    'eta': 0.1,
    'seed': 1000,
    'nthread': 4,
}

plst = params.items()
num_round = 10

bst = xgb.train(plst, dtrain, num_round)



y_pred = bst.predict(dtest)

score = accuracy_score(y_test, y_pred)

print("The accuracy score is: ", score)

Cross Validation

from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score


# CV model
model = xgb.XGBClassifier()
kfold = KFold(n_splits=4, random_state=666)
results = cross_val_score(model, X, y, cv=kfold)
print("Accuracy: %.2f%% (%.2f%%)" % (results.mean()*100, results.std()*100))

from sklearn.preprocessing import StandardScaler

ss = StandardScaler()
X_scaled = ss.fit_transform(X)
# CV model
model = xgb.XGBClassifier()
kfold = KFold(n_splits=4, random_state=666)
results = cross_val_score(model, X_scaled, y, cv=kfold)
print("Accuracy: %.2f%% (%.2f%%)" % (results.mean()*100, results.std()*100))



X_train_scaled = ss.fit_transform(X_train)
X_test_scaled = ss.transform(X_test)

# print(type(X_train_scaled))
# Xgboost manages only numeric vectors.
dtrain_scaled = xgb.DMatrix(X_train_scaled, label=y_train.values)
dtest_scaled = xgb.DMatrix(X_test_scaled, label=y_test.values)

bst2 = xgb.train(plst, dtrain_scaled, num_round)
y_pred2 = bst2.predict(dtest_scaled)



# print(type(y_pred2))

score2 = accuracy_score(y_test, y_pred2)

print("The accuracy score is: ", score2)

print(accuracy_score(y_pred, y_pred2))

print(y_pred2.shape)
print(y_test.shape)
print(y_pred.shape)
print(X_train_scaled.shape)
